fix(api): consolidate Wave 3.B + 3.C error mapping to domain exceptions
Replaces every remaining ClientException and manual Response(error=...)
construction with the domain-error hierarchy that flows through the
central RFC 9457 exception handler.

- Add FeatureNotImplementedError (501) + ErrorCode.FEATURE_NOT_IMPLEMENTED
  (8009) so the 501 'Fine-tuning is not available' /
  BackendUnsupportedError sites in memory.py can raise a typed domain
  error instead of ClientException(status_code=501).
- Migrate ClientException raises in memory, audit, coordination_metrics,
  custom_rules, settings to ConflictError / ValidationError /
  FeatureNotImplementedError / NotFoundError as appropriate.  Drop the
  litestar.exceptions.ClientException imports where no callers remain.
- Migrate manual Response(...) error returns in
  agent_identity_versions and artifacts to raise NotFoundError /
  ValidationError / InternalServerException.  download_content's
  pre-stream 404 paths raise rather than fabricating a JSON envelope
  inline; SSE error format in providers.pull_model gets a carve-out
  comment because headers are already on the wire by the time the
  generator yields.
- Wave 3.B follow-on: relocate provider-management DTOs and per-op
  rate-limit configs out of synthorg.api into the domain modules they
  belong to (synthorg.providers.management.dtos /
  capability_dtos, synthorg.config.rate_limits).  api.dto_providers,
  api.dto_provider_capabilities, and api.rate_limits.{config,
  inflight_config} now re-export the domain types so existing imports
  keep working without a behavioural shim.
- Update tests that asserted the legacy 400 status code or the
  ClientException type to expect the typed domain exceptions and the
  RFC 9457 422 status they now produce.
- Regenerate mock_spec_baseline.txt to track the line-number shifts
  from the import-block edits.
