+++ src/synthorg/api/controllers/memory.py
@@ -9,15 +9,18 @@ from typing import Final
 
 from litestar import Controller, delete, get, post
 from litestar.datastructures import State  # noqa: TC002
-from litestar.exceptions import ClientException
-from litestar.status_codes import HTTP_409_CONFLICT, HTTP_501_NOT_IMPLEMENTED
 from pydantic import BaseModel, ConfigDict, Field
 
 from synthorg.api.dto import ApiResponse, PaginatedResponse, PaginationMeta
 from synthorg.api.guards import HumanRole, require_roles
 from synthorg.api.rate_limits import per_op_concurrency, per_op_rate_limit_from_policy
 from synthorg.api.state import AppState  # noqa: TC001
-from synthorg.core.domain_errors import NotFoundError
+from synthorg.core.domain_errors import (
+    ConflictError,
+    FeatureNotImplementedError,
+    NotFoundError,
+    ValidationError,
+)
 from synthorg.core.persistence_errors import QueryError
 from synthorg.core.types import NotBlankStr
 from synthorg.memory.embedding.fine_tune import FineTuneStage
@@ -84,14 +87,15 @@ def _build_memory_service(
         app_state: Active application state.
         require_fine_tune: When ``True`` (default), eagerly resolve
             ``fine_tune_checkpoints`` / ``fine_tune_runs`` and raise
-            :class:`ClientException` (HTTP 501) when they are absent.
-            When ``False``, leave the repos as ``None`` so the service
-            constructs cleanly for memory-only endpoints.
+            :class:`FeatureNotImplementedError` (HTTP 501) when they
+            are absent.  When ``False``, leave the repos as ``None``
+            so the service constructs cleanly for memory-only
+            endpoints.
 
     Raises:
-        ClientException: When ``require_fine_tune`` is ``True`` and the
-            backend does not implement the fine-tune repositories
-            (HTTP 501).
+        FeatureNotImplementedError: When ``require_fine_tune`` is
+            ``True`` and the backend does not implement the fine-tune
+            repositories (HTTP 501).
     """
     backend = app_state.persistence
     checkpoint_repo: FineTuneCheckpointRepository | None = None
@@ -101,13 +105,11 @@ def _build_memory_service(
             checkpoint_repo = backend.fine_tune_checkpoints
             run_repo = backend.fine_tune_runs
         except NotImplementedError as exc:
-            raise ClientException(
-                detail=(
-                    "Fine-tune admin endpoints are not supported by the "
-                    "active persistence backend."
-                ),
-                status_code=HTTP_501_NOT_IMPLEMENTED,
-            ) from exc
+            msg = (
+                "Fine-tune admin endpoints are not supported by the "
+                "active persistence backend."
+            )
+            raise FeatureNotImplementedError(msg) from exc
     return MemoryService(
         checkpoint_repo=checkpoint_repo,
         run_repo=run_repo,
@@ -189,7 +191,8 @@ class MemoryAdminController(Controller):
             base_model=data.base_model,
         )
         if not app_state.has_fine_tune_orchestrator:
-            raise ClientException(detail="Fine-tuning is not available")
+            msg = "Fine-tuning is not available"
+            raise FeatureNotImplementedError(msg)
         orchestrator = app_state.fine_tune_orchestrator
         try:
             run = await orchestrator.start(data)
@@ -198,10 +201,8 @@ class MemoryAdminController(Controller):
                 MEMORY_FINE_TUNE_REQUESTED,
                 error=str(exc),
             )
-            raise ClientException(
-                detail="A fine-tuning run is already active",
-                status_code=HTTP_409_CONFLICT,
-            ) from exc
+            msg = "A fine-tuning run is already active"
+            raise ConflictError(msg) from exc
         return ApiResponse(
             data=FineTuneStatus(
                 run_id=run.id,
@@ -233,7 +234,8 @@ class MemoryAdminController(Controller):
         """Resume a failed/cancelled pipeline run."""
         app_state: AppState = state.app_state
         if not app_state.has_fine_tune_orchestrator:
-            raise ClientException(detail="Fine-tuning is not available")
+            msg = "Fine-tuning is not available"
+            raise FeatureNotImplementedError(msg)
         orchestrator = app_state.fine_tune_orchestrator
         try:
             run = await orchestrator.resume(run_id)
@@ -243,19 +245,16 @@ class MemoryAdminController(Controller):
                 run_id=run_id,
                 error=str(exc),
             )
-            raise ClientException(
-                detail="A fine-tuning run is already active",
-                status_code=HTTP_409_CONFLICT,
-            ) from exc
+            msg = "A fine-tuning run is already active"
+            raise ConflictError(msg) from exc
         except ValueError as exc:
             logger.warning(
                 MEMORY_FINE_TUNE_REQUESTED,
                 run_id=run_id,
                 error=str(exc),
             )
-            raise ClientException(
-                detail="Run not found or not resumable",
-            ) from exc
+            msg = "Run not found or not resumable"
+            raise NotFoundError(msg) from exc
         return ApiResponse(
             data=FineTuneStatus(
                 run_id=run.id,
@@ -292,7 +291,8 @@ class MemoryAdminController(Controller):
         """Cancel the active pipeline run."""
         app_state: AppState = state.app_state
         if not app_state.has_fine_tune_orchestrator:
-            raise ClientException(detail="Fine-tuning is not available")
+            msg = "Fine-tuning is not available"
+            raise FeatureNotImplementedError(msg)
         orchestrator = app_state.fine_tune_orchestrator
         await orchestrator.cancel()
         status = await orchestrator.get_status()
@@ -392,10 +392,8 @@ class MemoryAdminController(Controller):
         except CheckpointNotFoundError as exc:
             raise NotFoundError(str(exc)) from exc
         except QueryError as exc:
-            raise ClientException(
-                detail="Failed to update embedder settings",
-                status_code=HTTP_409_CONFLICT,
-            ) from exc
+            msg = "Failed to update embedder settings"
+            raise ConflictError(msg) from exc
         return ApiResponse(data=updated)
 
     @post(
@@ -421,10 +419,10 @@ class MemoryAdminController(Controller):
 
         Exception mapping:
 
-        - ``CheckpointNotFoundError`` -> HTTP 404
+        - ``CheckpointNotFoundError`` -> HTTP 404 via ``NotFoundError``
         - ``CheckpointRollbackUnavailableError``,
-          ``CheckpointRollbackCorruptError`` -> HTTP 400 via
-          ``ClientException`` (operator error / corrupt backup)
+          ``CheckpointRollbackCorruptError`` -> HTTP 422 via
+          ``ValidationError`` (operator error / corrupt backup)
         - Any other exception propagates as HTTP 500
         """
         service = _build_memory_service(state.app_state)
@@ -433,9 +431,11 @@ class MemoryAdminController(Controller):
         except CheckpointNotFoundError as exc:
             raise NotFoundError(str(exc)) from exc
         except CheckpointRollbackUnavailableError as exc:
-            raise ClientException(detail=str(exc)) from exc
+            # Operator-error / corrupt backup conditions; 422 better
+            # reflects "rollback target invalid" than a generic 400.
+            raise ValidationError(str(exc)) from exc
         except CheckpointRollbackCorruptError as exc:
-            raise ClientException(detail=str(exc)) from exc
+            raise ValidationError(str(exc)) from exc
         return ApiResponse(data=updated)
 
     @delete(
@@ -469,10 +469,7 @@ class MemoryAdminController(Controller):
         except CheckpointNotFoundError as exc:
             raise NotFoundError(str(exc)) from exc
         except QueryError as exc:
-            raise ClientException(
-                detail=str(exc),
-                status_code=HTTP_409_CONFLICT,
-            ) from exc
+            raise ConflictError(str(exc)) from exc
         return ApiResponse(data=None)
 
     # -- Memory entries (GDPR) ---------------------------------------
@@ -515,9 +512,8 @@ class MemoryAdminController(Controller):
             # ``MEMORY_ENTRY_DELETE_FAILED`` for this branch, so the
             # controller stays in the layering role of HTTP
             # translation only and does not double-record the event.
-            raise ClientException(
-                detail=str(exc),
-                status_code=HTTP_501_NOT_IMPLEMENTED,
+            raise FeatureNotImplementedError(
+                safe_error_description(exc),
             ) from exc
         if not deleted:
             # ``MemoryService.delete_memory_entry`` emits
