+++ src/synthorg/api/rate_limits/inflight_config.py
@@ -1,63 +1,10 @@
-"""Per-operation inflight-concurrency configuration."""
+"""Per-operation inflight-concurrency configuration (HTTP-facing alias).
 
-from typing import Literal, Self
+Source-of-truth lives in :mod:`synthorg.config.rate_limits` so the
+``synthorg.settings`` subsystem can consume it without crossing into
+the API layer (audit-144).
+"""
 
-from pydantic import BaseModel, ConfigDict, Field, model_validator
+from synthorg.config.rate_limits import PerOpConcurrencyConfig
 
-from synthorg.core.types import NotBlankStr  # noqa: TC001
-from synthorg.observability import get_logger
-from synthorg.observability.events.api import API_APP_STARTUP
-
-logger = get_logger(__name__)
-
-
-class PerOpConcurrencyConfig(BaseModel):
-    """Configuration for the per-operation inflight limiter.
-
-    Attributes:
-        enabled: Master switch.  When ``False`` the middleware becomes
-            a no-op and never attempts to acquire permits.
-        backend: Discriminator selecting the concrete
-            :class:`InflightStore` strategy.
-        overrides: Operator tuning knob.  Maps operation name to
-            ``max_inflight`` (positive integer) that supersedes the
-            decorator defaults.  Use ``0`` to explicitly disable an
-            operation (the middleware short-circuits and lets every
-            request through).  Negative values are invalid and rejected
-            at startup.
-    """
-
-    model_config = ConfigDict(frozen=True, allow_inf_nan=False)
-
-    enabled: bool = True
-    # Tighten to the only shipped backend.  A Redis adapter for
-    # cross-worker fairness is planned; adding it here must land with
-    # both the factory branch and the corresponding settings-enum
-    # entry in lockstep so an operator never picks a selectable value
-    # the factory raises ``NotImplementedError`` on at app construction.
-    backend: Literal["memory"] = "memory"
-    overrides: dict[NotBlankStr, int] = Field(default_factory=dict)
-
-    @model_validator(mode="after")
-    def _validate_override_values(self) -> Self:
-        """Reject negative override values.
-
-        Zero is allowed and means "disable this operation" -- the
-        middleware short-circuits when the override is ``0``.  Bad
-        configs are logged at WARNING before the ValueError is raised
-        so operator-facing config errors surface with context.
-        """
-        for operation, value in self.overrides.items():
-            if value < 0:
-                msg = (
-                    f"overrides[{operation!r}]={value!r} has negative value; "
-                    "use 0 to disable an operation"
-                )
-                logger.warning(
-                    API_APP_STARTUP,
-                    operation=operation,
-                    override=value,
-                    error=msg,
-                )
-                raise ValueError(msg)
-        return self
+__all__ = ["PerOpConcurrencyConfig"]
