+++ tests/unit/api/controllers/test_coordination_metrics.py
@@ -224,7 +224,7 @@ class TestCoordinationMetricsController:
         self,
         test_client: TestClient[Any],
     ) -> None:
-        """since > until returns 400."""
+        """since > until is a validation failure (HTTP 422)."""
         t1 = datetime(2026, 4, 1, tzinfo=UTC)
         t2 = t1 - timedelta(hours=1)
         resp = test_client.get(
@@ -232,4 +232,4 @@ class TestCoordinationMetricsController:
             params={"since": t1.isoformat(), "until": t2.isoformat()},
             headers=_HEADERS,
         )
-        assert resp.status_code == 400
+        assert resp.status_code == 422
